Added fix to TSS from boost CVS for static thread_specific_ptr. Fixed a couple tss leaks. Silenced WSA_OPERATION_ABORTED error when listening sockets are closed. Silenced warning with static_cast.
git-svn-id: svn+ssh://localhost/home/svn/musador/trunk@42 abf014a6-d86a-7740-8175-6a5112e866a6

+++ src/Server/test/ServerTest.h
@@ -138,7 +138,7 @@ public:
 		std::vector<char> echo(MAX);
 		while (sent < MAX)
 		{
-			int nBytes = 1 + r.rand(MAX - sent - 1);
+			int nBytes = 1 + static_cast<int>(r.rand(MAX - sent - 1));
 			
 			try
 			{
